dia9/etl-neoauto-paginado/requirements.txt
prefect
requests
bs4
mysql-connector-python